add interface for cloud + endpoints
