+++ src/Promitor.Agents.ResourceDiscovery/Configuration/AzureLandscape.cs
@@ -1,10 +1,11 @@
 ﻿using System.Collections.Generic;
 using Promitor.Core.Configuration;
+using Promitor.Core.Extensions;
 using Promitor.Core.Serialization.Enum;
 
 namespace Promitor.Agents.ResourceDiscovery.Configuration
 {
-    public class AzureLandscape
+    public class AzureLandscape : IAzureCloudEndpoints
     {
         public string TenantId { get; set; }
         public List<string> Subscriptions { get; set; }
