+++ src/Promitor.Core.Scraping/Configuration/Model/AzureMetadata.cs
@@ -1,9 +1,10 @@
 ﻿using Promitor.Core.Configuration;
+using Promitor.Core.Extensions;
 using Promitor.Core.Serialization.Enum;
 
 namespace Promitor.Core.Scraping.Configuration.Model
 {
-    public class AzureMetadata
+    public class AzureMetadata : IAzureCloudEndpoints
     {
         public string ResourceGroupName { get; set; }
         public string SubscriptionId { get; set; }
