+++ src/Promitor.Core/Extensions/IAzureCloudEndpoints.cs
@@ -1,0 +1,12 @@
+
+using Promitor.Core.Configuration;
+using Promitor.Core.Serialization.Enum;
+
+namespace Promitor.Core.Extensions
+{
+    public interface IAzureCloudEndpoints
+    {
+        AzureCloud Cloud { get; }
+        AzureEndpoints Endpoints { get; }
+    }
+}
